+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.2.0] - 2018-09-06
 ### Added
 - The `map` mutation.
 - The `makeCallbackMutation` mutation builder.
@@ -19,6 +21,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Added
 - The `filter` mutation for filtering a list of objects.
 
-[Unreleased]: https://github.com/CultureHQ/components/compare/v0.1.1...HEAD
-[0.1.0]: https://github.com/CultureHQ/components/compare/v0.1.0...v0.1.1
+[Unreleased]: https://github.com/CultureHQ/components/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/CultureHQ/components/compare/v0.1.1...v0.2.0
+[0.1.1]: https://github.com/CultureHQ/components/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/CultureHQ/components/compare/v0.0.4...v0.1.0